requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
Pygments==2.13.0
python-decouple==3.6
rich==12.6.0
shellingham==1.5.0
tomli==2.0.1
typer==0.6.1
